+++ concordium-consensus/src/Concordium/GlobalState/Account.hs
@@ -26,7 +26,6 @@ import Concordium.Constants
 import Concordium.Types.HashableTo
 import Concordium.GlobalState.Basic.BlockState.AccountReleaseSchedule
 
-
 -- |A list of credential IDs that have been removed from an account.
 data RemovedCredentials
     = EmptyRemovedCredentials
@@ -230,8 +229,10 @@ data EncryptedAmountUpdate =
 
 -- |An update to an account state.
 data AccountUpdate = AccountUpdate {
-  -- |Address of the affected account.
-  _auAddress :: !AccountAddress
+  -- |Index of the affected account.
+  _auIndex :: !AccountIndex
+  -- |Canonical address of the affected account.
+  ,_auAddress :: !AccountAddress
   -- |Optionally a new account nonce.
   ,_auNonce :: !(Maybe Nonce)
   -- |Optionally an update to the account amount.
@@ -243,8 +244,8 @@ data AccountUpdate = AccountUpdate {
 } deriving(Eq)
 makeLenses ''AccountUpdate
 
-emptyAccountUpdate :: AccountAddress -> AccountUpdate
-emptyAccountUpdate addr = AccountUpdate addr Nothing Nothing Nothing Nothing
+emptyAccountUpdate :: AccountIndex -> AccountAddress -> AccountUpdate
+emptyAccountUpdate ai addr = AccountUpdate ai addr Nothing Nothing Nothing Nothing
 
 updateAccountInformation :: AccountThreshold -> Map.Map CredentialIndex AccountCredential -> [CredentialIndex] -> AccountInformation -> AccountInformation
 updateAccountInformation threshold addCreds remove (AccountInformation oldCredKeys _) =